backend/requirements.txt
flask>=3.1.1
flask-cors>=3.0.10
openai>=1.93.3
python-dotenv>=1.1.1
sqlalchemy>=2.0.41
flask-sqlalchemy>=3.1.1
psycopg2-binary>=2.9.10
dependency-injector>=4.48.1
pydantic>=2.11.7
pydantic-to-typescript>=2.0.0
pydantic-xml>=2.17.3
gunicorn>=23.0.0
alembic>=1.16.4
lxml>=6.0.0
